Reorganized file structure
Now the tests as well as the game in the folder structure make more sense and are easier to navigate as the player object changes and factories were created. Deleted the no longer used AbstractBranchPlayerMethods to remove overlap between the Parallel and Serial types of branching players since it made things very complicated and wasn't worth the hassle.