Dev/cheat player (#8)
Implemented cheat player as well as broke up the player class. This included overhauling the organization of the project and creating factory objects as well. There are still many TODOs, some of which can be discarded and others which can be improved, but cheat player (along with the rest of the tests) work.