Created small class that can set heuristic for A*
This class for A* is like the CheatPlayer but it requires that the use does their own heuristic method. From an implementation perspective this could be MAYBE be better if this class that requires a heuristic was a super class of the CheatPlayer so that instead of changing the already set CheatPlayer heuristic we can have CheatPlayer just extend this class but with a specific set heurstic.