Set up scrabble heuristic/reorganized
I haven't tested if the scrabble heuristic has made actual improvements though or if it's working as intended. I need to make a test to see what percentage of a time a given cheat player can figure it out, and reorganize the cheat player so that it can take in a cost and heuristic and extensions of the cheat player can wrap specific heuristic and cost functions.