Two Letter Words Set Fixed/Scrabble Heuristic
TLDR: Scrabble heuristic better, two letters words set is because I added it that way than forgot I did that.

The first test I added was to grid to make sure two letter words are considered valid when checking if the grid is valid. This test passed, and since there were two letter words within branch forward test, this wasn't the issue.

An additional test was made in BranchEmpty and that was not the issue.

Turns out I had the scrabble words not including 2 letter words in these tests because it is a house rule I often use, and for the RunCheatPlayer I did the same because the DefaultWordsSet did not include two letter words. There MAY be a slight problem with A* when there isn't a solution but likely not, just something to keep track of. Because of this I quickly added a no two letter words set as well. Another issue is my scrabble library was kind of weird so it doesn't have qi or za which I've see used often in scrabble. The two letter words still weren't working though. After some small tests just by inspecting the results within the run cheat player the two letter words were working.

I didn't create a test to compare the new scrabble heuristic to old heuristic, but for the letters aabbeessttpolqu the scrabble heuristic took 5807 milliseconds while the other heuristic took 11986. This is because the scrabble heuristic prioritized playing the q first while it had more letters to work with. Additionally the scrabble heuristic will also be less likely to remove words that have rarer letters in them. A better heuristic involing the bounding box of the grid could probably be made as well.