Began benchmark test (need to fix players first)
Made a test to determine how often the current version of A* works (just the default size of hand heuristic, for now, becaue the organization of the AI players need to be changed in a later commit). The AIPlayerWrapper was abstract for essentially no reason. Since it is useful when you don't have a real life game and want to control the tiles you will get, so I made it concrete. Based on this I found I should probably create a variation that dumps it's hardest to use letter after a certain amount of time. Before that though I should probably reorganize so that the AIWrapper doesn't just have a default cost and heuristic but is instead is a class that takes cost a heuristic as parameters and extending classes of that can have specifci costs and heuristics. I also need to compare how the scrabble heuristic does compared to the normal one that is just the number of tiles in hand. Currently the benchmark isn't complete because it it's able to get through all of the 144 tiles in the hand by peeling. There are a lot of potential solutions, including dumping after a specific amount of time and implementing specific cost and heuristic functions. Instead of making a ton of extensions for AIPlayer, I think a static class of various costs and heuristics would make more sense. That way the scrabble heuristic used for running one of the players could be more cleanly done.

Doing these tests I find the solver in it's current state much more easily handles 30 tiles at once rather than 15 and getting peels from there, which makes sense. Unfortunately this isn't how the game works. Nonetheless, a AIPlayer in which it resets the whole board to solve again could be useful to have, as some other heuristics to test. Eventually after implementing the other player classes more effectively this will time how long it takes for various types of players to complete their tasks.