Cleaned/documented player classes
Refactored the player classes and made them work better together and a bit clearer to understand, as well as cleaning the TODOs I covered the classes in. The BranchingPlayerParallel and BranchingPlayerSerial have significant overlap, but are both just implementations of branching for the player, implementing AbstractBranching player, and possibly trying to refactor those copied sections would make thing more confusing. Nonetheless, it is a potential improvement within the players that wasn't touched for this commit. Some functions were moved that could be more generally applicable, but were in subclasses, particularly, playGrid was in the AbstractBranchingPlayer when it makes more sense to be within the Player class. This function was also changed to make it not play anything on the grid unless all the tiles needed were in the player's hand.