Added Scrabble Heuristic / CheatPlayerManualAStarWrapper bug fix
Added a few heuristic/cost functions to play with for the solver, as well as bug fix in which the CheatPlayerManualAStarWrapper class, no matter what heuristic or cost function was passed, would use the default ones because of the A* factory built before those were updated. RunCheatPlayer is now also configured so that it is easy to switch out heuristic, cost, and goal functions.