RunCheatPlayer with Forward Branching
With the RunCheatPlayer, using new Player class to keep track of the effective branching factor. Now using the ForwardBranchingPlayerSerial which has a "branch" function using A* that does not step backward (this can have the drawback of if many tiles are placed and the A* is starting from there, it is not capable of stepping back). For the purpose of a neural network heuristic function, this is a more streamlined approach.